requirements.txt
pyTelegramBotAPI==4.25.0
SQLAlchemy==2.0.40
matplotlib==3.10.1
python-dotenv==1.1.0

